# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "SalesERP API",
    "description": "Sales ERP Management System API",
    "version": "1.0.0"
  },
  "host": "erp-sales-backend.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/api/items/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "type": "string"
          },
          {
            "name": "propertyType",
            "in": "query",
            "type": "string"
          },
          {
            "name": "status",
            "in": "query",
            "type": "string"
          },
          {
            "name": "showInSales",
            "in": "query",
            "type": "string"
          },
          {
            "name": "isActive",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "branchId": {
                  "example": "any"
                },
                "itemCode": {
                  "example": "any"
                },
                "itemName": {
                  "example": "any"
                },
                "propertyType": {
                  "example": "any"
                },
                "projectName": {
                  "example": "any"
                },
                "location": {
                  "example": "any"
                },
                "area": {
                  "example": "any"
                },
                "dimensions": {
                  "example": "any"
                },
                "ratePerUnit": {
                  "example": "any"
                },
                "totalPrice": {
                  "example": "any"
                },
                "bookingAmount": {
                  "example": "any"
                },
                "status": {
                  "example": "any"
                },
                "floorNumber": {
                  "example": "any"
                },
                "totalFloors": {
                  "example": "any"
                },
                "bedrooms": {
                  "example": "any"
                },
                "bathrooms": {
                  "example": "any"
                },
                "parking": {
                  "example": "any"
                },
                "amenities": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "legalDetails": {
                  "example": "any"
                },
                "hsnCode": {
                  "example": "any"
                },
                "openingStock": {
                  "example": "any"
                },
                "currentStock": {
                  "example": "any"
                },
                "openingValue": {
                  "example": "any"
                },
                "showInSales": {
                  "example": "any"
                },
                "images": {
                  "example": "any"
                },
                "isActive": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/items/for-sales": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/items/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "itemCode": {
                  "example": "any"
                },
                "itemName": {
                  "example": "any"
                },
                "propertyType": {
                  "example": "any"
                },
                "projectName": {
                  "example": "any"
                },
                "location": {
                  "example": "any"
                },
                "area": {
                  "example": "any"
                },
                "dimensions": {
                  "example": "any"
                },
                "ratePerUnit": {
                  "example": "any"
                },
                "totalPrice": {
                  "example": "any"
                },
                "bookingAmount": {
                  "example": "any"
                },
                "status": {
                  "example": "any"
                },
                "floorNumber": {
                  "example": "any"
                },
                "totalFloors": {
                  "example": "any"
                },
                "bedrooms": {
                  "example": "any"
                },
                "bathrooms": {
                  "example": "any"
                },
                "parking": {
                  "example": "any"
                },
                "amenities": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "legalDetails": {
                  "example": "any"
                },
                "hsnCode": {
                  "example": "any"
                },
                "openingStock": {
                  "example": "any"
                },
                "currentStock": {
                  "example": "any"
                },
                "openingValue": {
                  "example": "any"
                },
                "showInSales": {
                  "example": "any"
                },
                "images": {
                  "example": "any"
                },
                "isActive": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/items/{id}/documents": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/items/{id}/documents/{documentId}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/items/{id}/status": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/items/{id}/toggle-active": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/items/soft-delete/{id}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/items/restore/{id}": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/items/deleted/list": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  }
}